Add Flask-Cache and refactor deprecated imports
