+++ kqueen_ui/wsgi.py
@@ -1,6 +1,6 @@
 """WGSI module to run application using Gunicorn server."""
 
-from kqueen_ui.server import app as application
+from kqueen_ui import app as application
 from werkzeug.contrib.fixers import ProxyFix
 
 if __name__ == '__main__':
