+++ requirements.txt
@@ -1,5 +1,6 @@
 Flask==0.12.2
 Flask-Babel==0.11.2
+Flask-Cache
 Flask-Mail
 Flask-WTF
 gunicorn
