+++ setup.py
@@ -20,6 +20,7 @@ setup(
     install_requires=[
         'Flask==0.12.2',
         'Flask-Babel==0.11.2',
+        'Flask-Cache',
         'Flask-WTF',
         'gunicorn',
         'python-dateutil',
